requirements.txt
customtkinter
Pillow
requests
python-dotenv
